+++ test/test_da4vid/pipeline/steps/test_rfdiffusion_step.py
@@ -4,7 +4,6 @@ import unittest
 import warnings
 
 import docker
-import dotenv
 
 from da4vid.gpus.cuda import CudaDeviceManager
 from da4vid.io import read_from_pdb
@@ -21,11 +20,13 @@ class RFdiffusionStepTest(unittest.TestCase):
     self.image = StaticConfig.get(DOTENV_FILE).rfdiffusion_image
     self.model_weights = StaticConfig.get(DOTENV_FILE).rfdiffusion_models_dir
     self.folder = os.path.join(RESOURCES_ROOT, 'steps_test', 'rfdiffusion_test', 'step_folder')
+    self.resume_folder = os.path.join(RESOURCES_ROOT, 'steps_test', 'rfdiffusion_test', 'rf_resume')
     self.client = docker.from_env()
     self.gpu_manager = CudaDeviceManager()
 
   def tearDown(self):
-    shutil.rmtree(self.folder)
+    if os.path.isdir(self.folder):
+      shutil.rmtree(self.folder)
     self.client.close()
 
   def test_rfdiffusion_step_with_one_sample(self):
@@ -52,7 +53,37 @@ class RFdiffusionStepTest(unittest.TestCase):
       gpu_manager=self.gpu_manager,
       config=config
     )
-    sample_set = step._execute(orig_set)
+    sample_set = step.execute(orig_set)
+    self.assertEqual(3, len(sample_set.samples()))
+    orig_sequence = orig_set.samples()[0].protein.sequence()
+    for sample in sample_set.samples():
+      self.assertEqual(orig_sequence[20:30], sample.protein.sequence()[20:30])
+
+  def test_rfdiffusion_resume(self):
+    pdb_demo = os.path.join(RESOURCES_ROOT, 'steps_test', 'rfdiffusion_test', 'demo.pdb')
+    orig_set = SampleSet()
+    orig_set.add_samples(Sample(
+      name='DEMO',
+      filepath=pdb_demo,
+      protein=read_from_pdb(pdb_demo)
+    ))
+    config = RFdiffusionStep.RFdiffusionConfig(
+      num_designs=3,
+      contacts_threshold=4,
+      rog_potential=11,
+      partial_T=5
+    )
+    step = RFdiffusionStep(
+      image=self.image,
+      name='rfdiffusion_demo',
+      folder=self.resume_folder,
+      epitope=Epitope('A', 21, 30),
+      model_dir=self.model_weights,
+      client=self.client,
+      gpu_manager=self.gpu_manager,
+      config=config
+    )
+    sample_set = step.resume(orig_set)
     self.assertEqual(3, len(sample_set.samples()))
     orig_sequence = orig_set.samples()[0].protein.sequence()
     for sample in sample_set.samples():
